# specs/001-github-project-chat/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: GitHub Projects Chat API
  description: |
    REST API for the GitHub Projects Chat Interface.
    Enables natural language task management for GitHub Projects V2.
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:8000/api/v1
    description: Development server
  - url: https://api.example.com/api/v1
    description: Production server
...
paths:
...
  /auth/github/callback:
    get:
      tags: [auth]
      summary: GitHub OAuth callback
      description: Handles OAuth callback, exchanges code for token
      operationId: handleGitHubCallback
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
        - name: state
          in: query
          required: true
          schema:
            type: string
      responses:
        '302':
          description: Redirect to frontend with session
        '400':
          $ref: '#/components/responses/BadRequest'
  /auth/me:
    get:
      tags: [auth]
      summary: Get current user
      description: Returns authenticated user information
      operationId: getCurrentUser
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Current user info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/logout:
    post:
      tags: [auth]
      summary: Logout user
      description: Invalidates the current session
      operationId: logout
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Successfully logged out
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Logged out successfully

  # ============ PROJECTS ENDPOINTS ============
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session_id
  schemas:
    User:
      type: object
      properties:
        github_user_id:
          type: string
        github_username:
          type: string
        github_avatar_url:
          type: string
          nullable: true
        selected_project_id:
          type: string
          nullable: true
...
    Error:
      type: object
      properties:
        error:
          type: string
        detail:
          type: string
        code:
          type: string
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Bad Request
            detail: Invalid project_id format
            code: INVALID_PROJECT_ID
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Unauthorized
            detail: Session expired or invalid
            code: UNAUTHORIZED
...